Optimizations to the EXISTS-annotated plan
A few minor tweaks, but the bulk of this commit revolves around making EXISTS-annotated
plans play nice with the scala optimizer.  There are two main fixes:
 - Caveats.strip / DataFrameImplicits.stripCaveats rewrites the plan to remove all
   ApplyCaveat operators.  This has no effect on the output of the plan, but will
   preclude the plan from being used for subsequent annotation operations.  In principle,
   since ApplyCaveat needs to be passed inputs for all of its attributes, this should
   make execution *slightly* faster.
 - CaveatExistsInExpression/CaveatExistsInPlan have been rewritten to use bare attributes
   instead of attributes assembled inside a struct.  The final step after the rewrite
   assembles the resulting attributes into the appropriate struct.

The latter optimization is due to a rule in the spark optimizer:
[CollapseProject](https://github.com/apache/spark/blob/1f0287148977adb416001cb0988e919a2698c8e0/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/optimizer/Optimizer.scala#L692)
This rule is simple projection inlining, but without optimization on the resulting
expression.  Consider the following trivial example with two projections:

```
Project[..., __MIMIR_CAVEAT <- BuildStruct(ROW <- __MIMIR_CAVEAT.ROW, ATTRIBUTES <- __MIMIR_CAVEAT.ATTRIBUTES)]
  Project[..., __MIMIR_CAVEAT <- BuildStruct(ROW <- AAA, ATTRIBUTES <- BBB)]
    ...
```

CollapseProject merges these operations together, inlining attributes, even if it involves
replicating them:
```
Project[...,
  __MIMIR_CAVEAT <- BuildStruct(
    ROW <- BuildStruct(ROW <- AAA, ATTRIBUTES <- BBB).ROW,
    ATTRIBUTES <- BuildStruct(ROW <- AAA, ATTRIBUTES <- BBB).ATTRIBUTES
  )]
    ...
```
This type of projection stacking is a common pattern in the EXISTS annotation rewrite,
since any manipulation of the annotation needs to disassemble and then reassemble the
annotation.  Unfortunately, `CollapseProject` ends up creating an exponential expression
in the number of this type of projections (even though the resulting expressions can
be optimized back down to a sane size).  To avoid this, we're now just using bare
attributes, which the optimizer knows how to deal with.