Filter predicates used in target slices for caveat enumeration (closes https://github.com/VizierDB/web-ui/issues/222)
As EnumerateCaveats(...) descends through a LogicalPlan, it tracks which portion
of the plan is "interesting" through something it calls a "slice": A predicate
defining the set of rows that contribute to the parts of the output that are
interesting for the query output.

With this commit, when EnumerateCaveats(...) encounters a Filter node, it now
incorporates the filter condition into all slices before descending further into
the plan.  In particular, this is what was causing too many caveats to show up in
Vizier (https://github.com/VizierDB/web-ui/issues/222)