Fixing an unresolved attribute bug in Aggregate annotation
So apparently analyzing a spark expression only guarantees you that the outermost plan will be analyzed.  Inner plans may or may not be analyzed.  In this particular case, that was happening, and exprIds weren't being properly generated for alias expressions.  I added a bit of code that should force the extraction of resolved attributes from aliases.